% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGBmonteCarlo_chave2014.R
\name{AGBmonteCarlo_chave2014}
\alias{AGBmonteCarlo_chave2014}
\title{Propagating above ground biomass (AGB) or carbon (AGC) errors to the stand level}
\usage{
AGBmonteCarlo_chave2014(
  D,
  WD = NULL,
  errWD = NULL,
  H = NULL,
  errH = NULL,
  HDmodel = NULL,
  coord = NULL,
  Dpropag = NULL,
  n = 1000,
  Dlim = NULL
)
}
\arguments{
\item{D}{Vector of tree diameters (in cm)}

\item{WD}{Vector of wood density estimates (in g/cm3)}

\item{errWD}{Vector of error associated to the wood density estimates (should be of the same size as \code{WD})}

\item{H}{(option 1) Vector of tree heights (in m). If set, \code{errH} must be set too.}

\item{errH}{(if \code{H}) Residual standard error (RSE) of a model or vector of errors (sd values) associated to tree height
values (in the latter case the vector should be of the same length as \code{H}).}

\item{HDmodel}{(option 2) Model used to estimate tree height from tree diameter (output from \code{\link[=modelHD]{modelHD()}}, see example).}

\item{coord}{(option 3) Coordinates of the site(s), either a vector giving a single site (e.g. c(longitude, latitude))
or a matrix/dataframe with two columns (e.g. cbind(longitude, latitude)). The coordinates are used to predict
height-diameter allometry with bioclimatic variables.}

\item{Dpropag}{This variable can take three kind of values, indicating how to propagate the errors on diameter measurements:
a single numerical value or a vector of the same size as \code{D}, both representing the standard deviation associated
with the diameter measurements or \code{"chave2004"} (an important error on 5 percent of the measures, a smaller error on
95 percent of the trees).}

\item{n}{Number of iterations. Cannot be smaller than 50 or larger than 1000. By default \code{n = 1000}}

\item{Dlim}{(optional) Minimum diameter (in cm) for which above ground biomass should be calculated (all diameter below
\code{Dlim} will have a 0 value in the output).}
}
\value{
Returns a matrix with the AGB of the trees (rows) times the n iterations (columns)
}
\description{
Propagation of the errors throughout the steps needed to compute AGB or AGC.
}
\details{
See Rejou-Mechain et al. (2017) for all details on the error propagation procedure.
}
\references{
Chave, J. et al. (2004). \emph{Error propagation and scaling for tropical forest biomass estimates}.
Philosophical Transactions of the Royal Society B: Biological Sciences, 359(1443), 409-420.

Rejou-Mechain et al. (2017).
\emph{BIOMASS: An R Package for estimating above-ground biomass and its uncertainty in tropical forests}.
Methods in Ecology and Evolution, 8 (9), 1163-1167.
}
\author{
Maxime REJOU-MECHAIN, Bruno HERAULT, Camille PIPONIOT, Ariane TANGUY, Arthur PERE
}
\keyword{Carlo}
\keyword{Monte}
\keyword{internal}
