% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{APG}
\alias{APG}
\title{APG dataset}
\format{
A data frame with 5766 observations on the following 3 variables :
\itemize{
\item \code{inputGenus}: Names of the genus
\item \code{family}: Names of the retrieved family
\item \code{order}: Names of the retrieved order
}
}
\usage{
data("APG")
}
\description{
Retrieved APG III families and orders from genus names of the sub-dataset KarnatakaForest using BIOMASS getTaxonomy function :
getTaxonomy(KarnatakaForestsub$genusCorr, findOrder = TRUE)
}
\references{
See "KarnatakaForest" dataset.
}
\keyword{datasets}
