% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeTree.R
\name{attributeTree}
\alias{attributeTree}
\title{Attribute the tree to the subplot}
\usage{
attributeTree(x, y, Plot, CoordAbs)
}
\arguments{
\item{x}{The x coordinate of the tree for each plot}

\item{y}{The y coordinate of the tree for axis}

\item{Plot}{The label of the plot (same length as x and y)}

\item{CoordAbs}{Output of the function cutTree}
}
\value{
A vector with the subplot for each trees
}
\description{
Attribute the tree to the subplot
}
\examples{




}
\author{
Arthur PERE
}
