% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilinearInterpolation.R
\name{bilinearInterpolation}
\alias{bilinearInterpolation}
\title{Bilinear interpolation of X and Y absolute coordinates}
\usage{
bilinearInterpolation(relCoord, cornerCoord, dimX, dimY)
}
\arguments{
\item{relCoord}{a matrix : the relative coordinates to transform}

\item{cornerCoord}{a data.table : the absolute corners coordinates and its numbering following a clockwise direction (names resp. "X","Y","cornerNum")}

\item{dimX}{a vector indicating the size of the plot on the X axis, in meters and in the relative coordinates system}

\item{dimY}{a vector indicating the size of the plot on the Y axis, in meters and in the relative coordinates system}
}
\value{
a data.table containing the absolute coordinates of desired points
}
\description{
Apply a bilinear interpolation to convert relative coordinates into absolute coordinates, using the 4 corners absolute coordinates of the plot.
}
\author{
Arthur Bailly
}
\keyword{Internal}
\keyword{bilinear}
\keyword{interpolation}
