% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOMASS-package.R
\docType{package}
\name{BIOMASS-package}
\alias{BIOMASS}
\alias{BIOMASS-package}
\title{BIOMASS: Estimating Aboveground Biomass and Its Uncertainty in Forests}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Contains functions for estimating above-ground biomass/carbon and its uncertainty in forests. These functions allow to (1) retrieve and correct taxonomy, (2) estimate wood density and its uncertainty, (3) build height-diameter models, (4) manage tree and plot coordinates, (5) estimate above-ground biomass/carbon at stand level with associated uncertainty, (6) predicting landscape maps of AGBD with associated uncertainty using LiDAR products. To cite ‘BIOMASS’, please use citation(‘BIOMASS’). For more information, see Réjou-Méchain et al. (2017) \doi{10.1111/2041-210X.12753}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://umr-amap.github.io/BIOMASS/}
  \item \url{https://github.com/umr-amap/BIOMASS/}
  \item Report bugs at \url{https://github.com/umr-amap/BIOMASS/issues/}
}

}
\author{
\strong{Maintainer}: Dominique Lamonica \email{dominique.lamonica@ird.fr}

Authors:
\itemize{
  \item Dominique Lamonica \email{dominique.lamonica@ird.fr}
  \item Maxime Réjou-Méchain \email{maxime.rejou@gmail.com} [data contributor]
  \item Arthur Bailly \email{arthur.bailly@ird.fr}
  \item Guillaume Cornu \email{gcornu@cirad.fr} (\href{https://orcid.org/0000-0002-7523-5176}{ORCID})
  \item Arthur Pere
  \item Ariane Tanguy
  \item Camille Piponiot
  \item Bruno Hérault \email{bruno.herault@cirad.fr}
}

Other contributors:
\itemize{
  \item John Godlee \email{GodleeJ@si.edu} [contributor]
  \item Fabian Fischer \email{f.fischer@bristol.ac.uk} [contributor]
  \item Jerome Chave \email{jerome.chave@univ-tlse3.fr} [contributor]
  \item Philippe Verley \email{philippe.verley@ird.fr} [contributor]
  \item Ted Feldpausch \email{T.R.Feldpausch@exeter.ac.uk} [data contributor]
}

}
\keyword{internal}
