% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheManager.R
\name{cacheManager}
\alias{cacheManager}
\title{Function that return a possibly cached file, transparently downloading it if missing}
\usage{
cacheManager(nameFile)
}
\arguments{
\item{nameFile}{character. file to resolve cached path.}
}
\value{
file path of the resolved cached file.
}
\description{
Function that return a possibly cached file, transparently downloading it if missing
}
\section{Localisation}{

Cache path discovery protocol
\enumerate{
\item BIOMASS.cache option set to an \strong{existing} folder
\item \strong{existing} user data folder \code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}}
\itemize{
\item On Linux : \verb{~/.local/share/R/BIOMASS}
\item On Mac OS X : \verb{~/Library/Application Support/R/BIOMASS}
\item On Windows 7 up to 10 : \verb{C:\\\\Users\\\\<username>\\\\AppData\\\\Local\\\\R\\\\BIOMASS}
\item On Windows XP : \verb{C:\\\\Documents and Settings\\\\<username>\\\\Data\\\\R\\\\BIOMASS}
}
\item fallback to R session tempdir
}
}

\author{
Guillaume CORNU
}
