% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheManager.R
\name{clearCache}
\alias{clearCache}
\title{Function to clear cache content and possibly remove it}
\usage{
clearCache(remove = FALSE)
}
\arguments{
\item{remove}{logical. If TRUE cache folder will be removed too (not only content)
resulting in deactivating cache as a side effect}
}
\value{
No return value, called for side effects
}
\description{
It will refuse to clear or remove a custom cache folder set using BIOMASS.cache
option as we don't know whether this folder contains other possibly valuable
files apart from our cached files.
}
