% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAGB.R
\name{computeAGB}
\alias{computeAGB}
\title{Computing tree above ground biomass (AGB)}
\usage{
computeAGB(allometry = chave2014, ...)
}
\arguments{
\item{allometry}{a function which returns AGB estimates as a numeric vector}

\item{...}{arguments passed to the function provided in \code{allometry}}
}
\value{
numeric vector with tree AGB estimates
}
\description{
Compute tree above ground biomass (AGB) using multiple biomass allometry
NOT A FAN OF MULTIPLE ALLOMETRY, NOT HOMOGENEOUS WITH THE REST OF AGB COMPUTATION
}
\examples{
allom <- function(D, G) {
  D*G
}
allom2 <- function(G) {
  G^2
}

d <- c(10:99)
g <- c(100:189)

computeAGB(allometry = allom, D = d, G = g)
computeAGB(allometry = allom2, G = g)

}
