% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAGB_myallom.R
\name{computeAGB_myallom}
\alias{computeAGB_myallom}
\title{Compute tree above ground biomass (AGB) using (multiple) user biomass allometry}
\usage{
computeAGB_myallom(allometry, ...)
}
\arguments{
\item{allometry}{a function which returns AGB estimates as a numeric vector, or a list of functions}

\item{...}{arguments passed to function(s) provided in \code{allometry}}
}
\value{
numeric vector with tree AGB estimates
}
\description{
Compute tree above ground biomass (AGB) using (multiple) user biomass allometry
}
\examples{
allom <- function(D, G) {
  D*G
}
allom2 <- function(G) {
  G^2
}

d <- c(10:99)
g <- c(100:189)

computeAGB_myallom(list(allom, allom2), D = d, G = g)

}
