% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFeldRegion.R
\name{computeFeldRegion}
\alias{computeFeldRegion}
\title{Retriving the Feldpausch Region}
\usage{
computeFeldRegion(coord)
}
\arguments{
\item{coord}{Coordinates of the site(s), a matrix/dataframe with two columns (e.g. cbind(longitude, latitude)) (see examples).}
}
\value{
The function return a vector with The Feldpausch Region prepared for the \code{retrieveH} function
}
\description{
Extract the Feldpausch et al. (2012) Region thanks to the coordinates of the data. 
The function is time-consuming at its first use as it downloads a raster in the folder FeldRegion (see details).
However, as soon as the raster is downloaded once, the function then runs fast.
}
\section{Localisation}{

The localisation of this folder is :
\itemize{
\item On Linux : \code{~/.local/share/BIOMASS}
\item On Mac OS X : \code{~/Library/Application Support/BIOMASS}
\item On Windows 7 up to 10 : \code{C:\\Users\\<username>\\AppData\\Local\\BIOMASS\\BIOMASS}
\item On Windows XP : \code{C:\\Documents and Settings\\<username>\\Data\\BIOMASS\\BIOMASS}
}
See this function for more information : \code{\link[rappdirs]{user_data_dir}}
}

\examples{
#' # One study site 
lat <- 4.08 
long <- -52.68 
coord <- cbind(long, lat)
\dontrun{FeldRegion <- computeFeldRegion(coord)}

# Several study sites (here three sites)
long <- c(-52.68, -51.12, -53.11)
lat <- c(4.08, 3.98, 4.12)
coord <- cbind(long, lat)
\dontrun{FeldRegion <- computeFeldRegion(coord)}



}
\references{
Feldpausch et al. \emph{Tree height integrated into pantropical forest biomass estimates.} Biogeosciences (2012): 3381-3403.
}
\author{
Maxime REJOU-MECHAIN, Arthur PERE
}
