% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctCoordGPS.R
\name{correctCoordGPS}
\alias{correctCoordGPS}
\title{Correct the GPS coordinate}
\usage{
correctCoordGPS(longlat = NULL, UTMcoord = NULL, CoordRel, rangeX,
  rangeY, MaxDist = 10, drawPlot = F, RmOutliers = F)
}
\arguments{
\item{longlat}{(optionnal) data frame with the coordinate in longitude latitude (eg. cbind(longitude, latitude)).}

\item{UTMcoord}{(optionnal) data frame with the UTM coordinate in X Y}

\item{CoordRel}{data frame with the relative coordinate in the same order than the longlat or UTMcoord}

\item{rangeX}{a vector of length equal 2 with the min of the range for X}

\item{rangeY}{a vector of length equal 2 with the min of the range for Y}

\item{MaxDist}{a double with the maximum for the distance you want to use to make outliers}

\item{drawPlot}{a logical if you want to draw the plot}

\item{RmOutliers}{a logical if you want to remove the outliers}
}
\value{
If you haven't outliers or RmOutliers = TRUE, there will be a list with :
\describe{
   \item{corner}{a matrix with the coordinate of the corners whith no order in particular}
   \item{polygon}{a spatial polygone}
}
However if you have outliers and RmOutliers = FALSE, there will be data frame with :
\describe{
   \item{outlayers} {the line of your longlat or UTMcoord data frame, where it's an outlayers}
   \item{X} {the UTM coordinate X outlayer}
   \item{Y} {the UTM coordinate X outlayer}
}
}
\description{
If you have multiple GPS coordinate for one plot, you can use this function to correct them and take the corner of the plot
with the polygon associated.
}
\details{
You must give one parameter between longlat and UTMcoord
}
\examples{
UTMcoord = data.frame(X = c( runif(5, min = 9, max = 11), runif(5, min = 8, max = 12), runif(5, min = 80, max = 120), runif(5, min = 90, max = 110)),
                      Y = c( runif(5, min = 9, max = 11), runif(5, min = 80, max = 120), runif(5, min = 8, max = 12), runif(5, min = 90, max = 110)))
UTMcoord = UTMcoord + 1000
CoordRel = data.frame(X = c(rep(0, 10), rep(100, 10)),
                     Y = c(rep( c(rep(0, 5), rep(100, 5)), 2)))

aa=correctCoordGPS(UTMcoord = UTMcoord, CoordRel = CoordRel, rangeX = c(0, 100), rangeY = c(0, 100), drawPlot = T)
bb=correctCoordGPS(UTMcoord = UTMcoord, CoordRel = CoordRel, rangeX = c(0, 100), rangeY = c(0, 100), drawPlot = T,RmOutliers = T)
}
\author{
Arthur PERE, Maxime REJOU-MECHAIN

Arthur PERE, Maxime REJOU-MECHAIN
}
