% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctTaxo.R
\name{correctTaxo}
\alias{correctTaxo}
\title{Correct and match taxonomic names to the World Flora Online database}
\usage{
correctTaxo(
  genus,
  species = NULL,
  interactive = TRUE,
  preferAccepted = FALSE,
  preferFuzzy = FALSE,
  sub_pattern = subPattern(),
  useCache = FALSE,
  useAPI = TRUE,
  capacity = 60,
  fill_time_s = 60,
  timeout = 10
)
}
\arguments{
\item{genus}{vector of genera. Alternatively, the whole taxonomic name
(genus + species)}

\item{species}{optional, vector of species epithets to be checked (same
length as \code{genus})}

\item{interactive}{logical, if TRUE (default) user will be prompted to pick
names from a list where multiple ambiguous matches are found, otherwise
names with multiple ambiguous matches will be skipped}

\item{preferAccepted}{logical, if TRUE, if multiple ambiguous matches are
found, and if only one candidate is an "accepted" name,
automatically choose that name}

\item{preferFuzzy}{logical, if TRUE, if multiple ambiguous matches are
found, the accepted matched name with the lowest Levenshtein distance to
the submitted name will be returned}

\item{sub_pattern}{character vector of regex patterns which will be removed
from \code{paste(genus, species)} using \code{gsub()}. The order of this vector
matters, substitutions are applied sequentially. Sensible defaults are
provided by \code{subPattern()}}

\item{useCache}{logical, if TRUE use cached values in \code{the$wfo_cache}
preferentially, to reduce the number of API calls}

\item{useAPI}{logical, if TRUE (default) allow API calls}

\item{capacity}{maximum number of API calls which can accumulate over the
duration of \code{fill_time_s}. See documentation for \code{httr2::req_throttle()}}

\item{fill_time_s}{time in seconds to refill the capacity for repeated API
calls. See documentation for \code{httr2::req_throttle()}}

\item{timeout}{time in seconds to wait before disconnecting from an
unresponsive request}
}
\value{
data.frame of taxonomic names with rows matching \code{genus} + \code{species}.
\describe{
\item{nameOriginal}{Original name as in \code{genus} + \code{species}}
\item{nameSubmitted}{Name after optional sanitisation according to
\code{sub_pattern}}
\item{nameMatched}{Matched taxonomic name}
\item{nameAccepted}{Accepted taxonomic name}
\item{familyAccepted}{Family of accepted name}
\item{genusAccepted}{Genus of accepted name}
\item{speciesAccepted}{Species epithet of accepted name}
\item{nameModified}{Flag indicating if \code{matchedName} is different from
\code{nameOriginal}, not including the removal of excess whitespace}
}
}
\description{
Match taxonomic names using the World Flora Online database, via their
GraphQL API
}
\examples{
\donttest{
correctTaxo(genus = "Astrocarium", species = "standleanum")
correctTaxo(genus = "Astrocarium", species = "standleanum", interactive = F, preferFuzzy = T)
correctTaxo(genus = "Astrocarium standleanum", interactive = F, preferFuzzy = T)
}

}
\references{
Borsch, T. et al. (2020). \emph{World Flora Online: Placing taxonomists at the
heart of a definitive and comprehensive global resource on the world's
plants}. TAXON, 69, 6. doi10.1002/tax.12373:
}
\author{
John L. Godlee
}
