% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheManager.R
\name{createCache}
\alias{createCache}
\title{Function used to create or activate a permanent cache.}
\usage{
createCache(path = NULL)
}
\arguments{
\item{path}{Use a custom path to host cache}
}
\description{
Permanent cache is located by default in user data dir.
}
\details{
You can provide a custom path (that will be defined as a BIOMASS.cache option)
but clearCache function will refuse to operate on it for security reasons.
}
