% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{feldCoef}
\alias{feldCoef}
\title{Feldpausch et al. 2012 coefficients for generalized height-diameter models}
\format{A data frame with 12 observations on the following 4 variables: 
\describe{ 
  \item{\code{a}}{Coefficient a}
  \item{\code{b}}{Coefficient b}
  \item{\code{c}}{Coefficient c}
  \item{\code{RSE}}{Vector of RSE}
}}
\usage{
data("feldCoef")
}
\description{
Weibull coefficients from the Height-Diameter model \eqn{H = a(1-exp(-b*D^c))} by Feldpausch
et al. 2012., with the associated RSE.
}
\details{
This dataset is used in the function \code{\link{retrieveH}}, 
to predict height from diameter depending on the region.
}
\examples{
data(feldCoef)
str(feldCoef)
}
\references{
Feldpausch, Ted R., et al. \emph{Tree height integrated into pantropical forest biomass estimates.} Biogeosciences (2012): 3381-3403.
}
\keyword{datasets}
\keyword{internal}
