% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{feldRegion}
\alias{feldRegion}
\title{Raster with the region of Feldpausch et al. 2012 generalized height-diameter models}
\format{raster}
\usage{
data("feldCoef")
}
\description{
Raster with the region of Feldpausch et al. 2012 generalized height-diameter models
}
\details{
This dataset is used in the function \code{\link{retrieveH}}
to predict height from diameter depending on the region.
Feldpausch, T.R., et al. (2012). \emph{Tree height integrated into pantropical forest biomass estimates.} Biogeosciences, 9, 3381–3403.
}
\examples{
data(feldRegion)
}
\keyword{internal}
\keyword{raster}
