% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{genusFamily}
\alias{genusFamily}
\title{Genus Family database}
\format{A data frame with 28107 observations on the following 2 variables:
\itemize{
\item \code{family}: Vector of families APGIII corrected
\item \code{genus}: Vector of genus
}}
\source{
WCSP (2015). \emph{World Checklist of Selected Plant Families}.
Facilitated by the Royal Botanic Gardens, Kew. Published on the Internet; http://apps.kew.org/wcsp/ Retrieved 2015-12-17.

The Plant List (2013). Version 1.1. Published on the Internet; http://www.theplantlist.org/ Retrieved 2016-08-25.
}
\usage{
data("genusFamily")
}
\description{
To create this database, we combined the genera from The Plant List (http://www.theplantlist.org/1.1/browse/-/-/)
and the Vascular Plant Families and Genera from Kew (http://data.kew.org/vpfg1992/genlist.html).
Families were checked against the APGIII families.
}
\examples{
data(genusFamily)
str(genusFamily)
}
\keyword{datasets}
\keyword{internal}
