% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxonomy.R
\name{getTaxonomy}
\alias{getTaxonomy}
\title{Retrieve trees taxonomy}
\usage{
getTaxonomy(genus, findOrder = FALSE)
}
\arguments{
\item{genus}{Vector of genus names}

\item{findOrder}{(Boolean) If \code{TRUE}, the output will contain the taxonomical orders of the families.}
}
\value{
Data frame with the order (if \code{findOrder} is \code{TRUE}), family and genus.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{getTaxonomy()} is now deprecated, \code{correctTaxo()} being able to handle retrieving taxonomy all at once.

From given genus, the function finds the APG III family, and optionally the order, from the \link{genusFamily} database and the \link{apgFamilies} dataset
}
\examples{
# Find the Family of the Aphelandra genus
getTaxonomy("Aphelandra")
# ... and the order
\donttest{
getTaxonomy("Aphelandra", findOrder = TRUE)
}
}
\author{
Ariane TANGUY, Arthur PERE, Maxime REJOU-MECHAIN
}
