% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R, R/loglogFunction.R,
%   R/michaelisFunction.R, R/weibullFunction.R
\name{HDmethods}
\alias{HDmethods}
\alias{loglogFunction}
\alias{michaelisFunction}
\alias{weibullFunction}
\title{HDmethods}
\usage{
loglogFunction(
  data,
  weight = NULL,
  method,
  bayesian,
  useCache,
  chains,
  thin,
  iter,
  warmup,
  ...
)

michaelisFunction(
  data,
  weight = NULL,
  bayesian,
  useCache,
  chains,
  thin,
  iter,
  warmup,
  ...
)

weibullFunction(
  data,
  weight = NULL,
  bayesian,
  useCache,
  chains,
  thin,
  iter,
  warmup,
  ...
)
}
\arguments{
\item{data}{Dataset with the informations of height (H) and diameter (D)}

\item{weight}{(optional) Vector indicating observation weights in the model.}

\item{method}{In the case of the loglogFunction, the model is to be chosen between log1, log2 or log3.}

\item{bayesian}{a logical. If FALSE (by default) the model is estimated using a frequentist framework (lm or nls). If TRUE, the model is estimated in a Bayesian framework using the brms package.}

\item{useCache}{a logical. If bayesian = TRUE, determine wether to use the cache when building a Bayesian model (see Details).}

\item{chains}{(only relevant if bayesian = TRUE): Number of Markov chains (defaults to 3), see \code{\link[brms:brm]{brms::brm()}}}

\item{thin}{(only relevant if bayesian = TRUE): Thinning rate, see \code{\link[brms:brm]{brms::brm()}}}

\item{iter}{(only relevant if bayesian = TRUE): number of total iterations per chain (including warmup; defaults to 5000), see \code{\link[brms:brm]{brms::brm()}}}

\item{warmup}{(only relevant if bayesian = TRUE): number of warmup (aka burnin) iterations (defaults to 1000), see \code{\link[brms:brm]{brms::brm()}}}

\item{...}{Further arguments passed to \code{brm()}, e.g: prior, cores, etc. See \code{\link[brms:brm]{brms::brm()}}}
}
\value{
All the functions give an output similar to the one given by \code{\link[stats:lm]{stats::lm()}}, obtained for
\code{michaelisFunction} and \code{weibullFunction} from \link[minpack.lm:nlsLM]{minpack.lm::nlsLM}).

Result of a model (lm object if bayesian = FALSE, brm object if bayesian = TRUE)

Result of a model (nlsM object if bayesian = FALSE, brm object if bayesian = TRUE)

Result of a model (nlsM object if bayesian = FALSE, brm object if bayesian = TRUE)
}
\description{
Methods used for modeling height-diameter relationship
}
\details{
These functions model the relationship between tree height (H) and diameter (D).
\strong{loglogFunction}
Compute two types of log model (log and log2) to predict H from D.
The model can be:
\itemize{
\item log 1: \eqn{log(H) = a+ b*log(D)} (equivalent to a power model)
\item log 2: \eqn{log(H) = a+ b*log(D) + c*log(D)^2}
}

\strong{michaelisFunction}
Construct a Michaelis Menten model of the form: \deqn{H = (A * D) / (B + D)} (A and B are the model parameters to be estimated)

\strong{weibullFunction}
Construct a three parameter Weibull model of the form: \deqn{H = a*(1-exp(-(D/b)^c))} (a, b, c are the model parameters to be estimated)
}
\references{
Michaelis, L., & Menten, M. L. (1913). \emph{Die kinetik der invertinwirkung}. Biochem. z, 49(333-369), 352.
Weibull, W. (1951). \emph{Wide applicability}. Journal of applied mechanics, 103.
Baskerville, G. L. (1972). \emph{Use of logarithmic regression in the estimation of plant biomass}.
Canadian Journal of Forest Research, 2(1), 49-53.
}
\seealso{
\code{\link[=modelHD]{modelHD()}}
}
\author{
Maxime REJOU-MECHAIN, Ariane TANGUY
}
\keyword{Internal}
