% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong2UTM.R
\name{latlong2UTM}
\alias{latlong2UTM}
\title{Translate the long lat coordinate in UTM coordinate}
\usage{
latlong2UTM(Coord)
}
\value{
a data frame whith :
\describe{
   \item[long] The longitude of the entry
   \item[lat] The latitude of the entry
   \item[codeUTM] The code proj4 for UTM
   \item[X] The X UTM coordinate
   \item[Y] The Y UTM coordinate
}
}
\description{
Translate the long lat coordinate in UTM coordinate
}
