% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michaelisFunction.R
\name{michaelisFunction}
\alias{michaelisFunction}
\title{Modeling height-diameter relationship with Michaelis-Menten function}
\usage{
michaelisFunction(data, weight = NULL)
}
\arguments{
\item{data}{Dataset with the informations of height (H) and diameter (D)}

\item{weight}{(optional) Vector indicating observation weights in the model.}
}
\value{
This function give an output similar to the one given by \code{\link{lm}}, obtained for \code{michaelisFunction} from \code{\link[minpack.lm]{nlsLM}}).
}
\description{
Construct a Michaelis Menten model of the form: \deqn{H = (A * D) / (B + D)} (A and B are the model parameters to be estimated)
}
\references{
Michaelis, L., & Menten, M. L. (1913). \emph{Die kinetik der invertinwirkung}. Biochem. z, 49(333-369), 352.
}
\author{
Maxime REJOU-MECHAIN, Ariane TANGUY
}
