% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{NouraguesHD}
\alias{NouraguesHD}
\title{Height-Diameter data}
\format{A data frame with 1051 observations on the following variables :
\describe{
  \item{\code{plotId}}{Names of the plots}
  \item{\code{genus}}{Genus}
  \item{\code{species}}{Species}
  \item{\code{D}}{Diameter}
  \item{\code{H}}{Height}
  \item{\code{lat}}{Latitude}
  \item{\code{long}}{Longitude}
}}
\usage{
data("NouraguesHD")
}
\description{
Dataset from two 1-ha plots from the Nouragues forest (French Guiana)
}
\examples{
data(NouraguesHD)
str(NouraguesHD)
}
\references{
Ramesh, B. R. et al. (2010). \emph{Forest stand structure and composition in 96 sites along environmental 
gradients in the central Western Ghats of India} Ecological Archives E091-216. Ecology, 91(10), 3118-3118.
}
\keyword{datasets}
