% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{NouraguesPlot201}
\alias{NouraguesPlot201}
\title{Nouragues plot 201 coordinates}
\format{
A data frame with 40 (simulated GPS measurements) of the 8 following variables :
\itemize{
\item \code{Site}: Name of the site set up in the Nouragues forest
\item \code{Plot}: Plot ID of the site
\item \code{Xfield}: Corner location on the x-axis in the local coordinate system (defined by the 4 corners of the plot)
\item \code{Yfield}: Corner location on the y-axis in the local coordinate system
\item \code{Xutm}: Corner location on the x-axis in the UTM coordinate system
\item \code{Yutm}: Corner location on the y-axis in the UTM coordinate system
\item \code{Long}: Corner longitude coordinate
\item \code{Lat}: Corner latitude coordinate
}
}
\usage{
data(NouraguesPlot201)
}
\description{
Simulated corner coordinates of Nouragues 'Petit plateau' plot 201. The original coordinates have been modified to make the plot non-squared, and 10 repeated measurements of each corner have been simulated adding a random error to x and y coordinates.
}
\examples{
data(NouraguesPlot201)
str(NouraguesPlot201)
}
\references{
Jaouen, Gaëlle, 2023, "Nouragues forest permanent plots details", \doi{10.18167/DVN1/HXKS4E}, CIRAD Dataverse, V2
}
\keyword{datasets}
