% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{NouraguesTrees}
\alias{NouraguesTrees}
\title{Nouragues forest dataset}
\format{
A data frame with 2050 observations (trees) of the 8 following variables :
\itemize{
\item \code{site}: Name of the site set up in the Nouragues forest
\item \code{plot}: Plot ID
\item \code{Xfield}: Tree location on the x-axis in the local coordinate system (defined by the 4 corners of the plot)
\item \code{Yfield}: Tree location on the y-axis in the local coordinate system
\item \code{family}: Tree family
\item \code{genus}: Tree genus
\item \code{species}: Tree species
\item \code{D}: Tree diameter (in cm)
}
}
\usage{
data(NouraguesTrees)
}
\description{
This dataset contains 4 of the 12 plots of `Petit Plateau' permanent plots fifth census, 2012, Nouragues forestTree dataset (French Guiana). For educational purposes, some virtual trees have been added in the dataset. Dead trees have been removed.
}
\examples{
data(NouraguesTrees)
str(NouraguesTrees)
}
\references{
`Petit Plateau' permanent plots fifth census, 2012, Nouragues forest, \href{https://dataverse.cirad.fr/dataset.xhtml?persistentId=doi:10.18167/DVN1/TZ1RL9}{https://doi.org/10.18167/DVN1/TZ1RL9}, CIRAD Dataverse, V1
}
\keyword{datasets}
