% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCache.R
\name{removeCache}
\alias{removeCache}
\title{Reset all data stored in the cache}
\usage{
removeCache()
}
\description{
This function removes all the files that are automatically stored in a local folder following the use of
the correctTaxo, getBioclim or computeE functions.
}
\details{
This function first identifies the path where the cache is stored and then remove it
}
\examples{
removeCache()
}
\author{
Maxime REJOU-MECHAIN
}
