% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoryControl.R
\name{repertoryControl}
\alias{repertoryControl}
\title{Check if we have the file or repertory}
\usage{
repertoryControl(nameFile = "", correctTaxo = FALSE)
}
\arguments{
\item{nameFile}{the name of the file or repertory}

\item{correctTaxo}{(binary) if we are in the correctTaxo function}
}
\value{
the path to the repertory, and the separator if correctTaxo is False
}
\description{
Create a repertory if needed of it, control all of the file which needed to be downloaded and 
place in the specifique repertory
}
