% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctTaxo.R
\name{subPattern}
\alias{subPattern}
\title{Common taxonomic name substitutions}
\usage{
subPattern()
}
\value{
character vector with regular expressions for use with \code{gsub()}
}
\description{
Common taxonomic name substitutions
}
\details{
Used in argument \code{sub_pattern} in \code{correctTaxo()}
}
