% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryByPlot.R
\name{summaryByPlot}
\alias{summaryByPlot}
\title{Mixing the tree}
\usage{
summaryByPlot(plot, subplot = plot, AGB_simu)
}
\arguments{
\item{plot}{Vector with the code of plot}

\item{subplot}{vector with the code of the subplot}

\item{AGB_simu}{Matrix which result for the function \code{\link{AGBmonteCarlo}} on the AGB_simu element of the list}
}
\value{
The matrix corrected with the values of the trees who doesn't have a subplot replaced by NA
}
\description{
This function mixes the AGB of the tree that doesn't have a subplot assigned.
The AGB for those trees will be added with a random tree (in the same plot or
if the tree haven't a plot in the matrix) on each column of the matrix.
}
