% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{Taxo}
\alias{Taxo}
\title{Taxo dataset}
\format{
A data frame with 5766 observations on the following 3 variables :
\itemize{
\item \code{genusCorrected}: Corrected names of the genus
\item \code{speciesCorrected}: Corrected names of the species
\item \code{nameModified}: TRUE when names have been modified, FALSE if not
}
}
\usage{
data("Taxo")
}
\description{
Corrected taxonomy  of the sub-dataset KarnatakaForest using BIOMASS correctTaxo function :
correctTaxo(genus = KarnatakaForestsub$genus, species = KarnatakaForestsub$species, useCache = FALSE, verbose = FALSE)
}
\references{
See "KarnatakaForest" dataset.
}
\keyword{datasets}
