% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheManager.R
\name{updateCache}
\alias{updateCache}
\title{Update the cache for the different function}
\usage{
updateCache(nameFile = NULL)
}
\arguments{
\item{nameFile}{The name of the file you want to update. If it's \code{NULL} the function will update all the files.}
}
\value{
No return value, called for side effects
}
\description{
This function update the cache for the environmental variables:
\itemize{
\item wc2-5
\item CWD
\item E
}
}
\examples{
\donttest{
updateCache()
}
}
\author{
Arthur PERE
}
