% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{wsg_estimates}
\alias{wsg_estimates}
\title{The global wood density database}
\format{
A data frame with 20332 observations on the following 6 variables.
\itemize{
\item \code{family}: a character vector indicating the family
\item \code{genus}: a character vector indicating the genus
\item \code{species}: a character vector indicating the species
\item \code{wsg}: a numeric vector of mean wood density of a population (g/cm^3) at taxonomic level
\item \code{sd}: a numeric vector of standard deviation of wood density of a popuation (g/cm^3) at taxonomic level
\item \code{level_tax}: a character vector of taxonomic level (namely \code{family} or \code{genus} or \code{species})
}
}
\source{
Fischer, F. J., et al. (2026). Global Wood Density Database v.2 (GWDD v.2) (Data set). Zenodo. https://doi.org/10.5281/zenodo.18262736
}
\usage{
data("wsg_estimates")
}
\description{
The Global Wood Density Database v.2 (GWDD v.2) (Fischer, F. J., et al. 2026).
}
\details{
This dataset is used in the function \code{\link[=getWoodDensity]{getWoodDensity()}}, to estimate a taxon-average wood density value.
}
\examples{
data(wsg_estimates)
str(wsg_estimates)
}
\references{
Fischer, F. J., et al. (2026). Beyond species means - the intraspecific contribution to global wood density variation. New Phytol. https://doi.org/10.1111/nph.70860
}
\keyword{datasets}
\keyword{density}
\keyword{getWoodDensity}
\keyword{internal}
\keyword{wood}
