% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeTreeCoord.R
\name{attributeTreeCoord}
\alias{attributeTreeCoord}
\title{Attribute trees to GPS coordinates}
\usage{
attributeTreeCoord(xy, plot, coordGPS = NULL, coordRel = NULL,
  plotCoord = NULL, coordAbs = NULL)
}
\arguments{
\item{xy}{The coordinates of the trees for each plot}

\item{plot}{The label of the plot (same length as the number of rows of \code{xy})}

\item{coordGPS}{(optional) The GPS coordinate of the plot}

\item{coordRel}{(optional) The relative coordinate for the GPS points of the plot}

\item{plotCoord}{(optional) The name of the plot for the differents GPS points}

\item{coordAbs}{(optional) The result of the function \code{\link[=cutPlot]{cutPlot()}}}
}
\value{
A data frame with two columns :
- \code{X}: The \code{X} coordinate in whichever projection you have
- \code{Y}: The \code{X} coordinate in whichever projection you have
}
\description{
Attribute trees to GPS coordinates
}
\examples{

# Trees relative coordinates
xy <- data.frame(x = runif(200, min = 0, max = 200), y = runif(200, min = 0, max = 200))


# cut the plot in multiple part
coord <- data.frame(X = rep(c(0, 200, 0, 200), 2), Y = rep(c(0, 0, 200, 200), 2))
coord[1:4, ] <- coord[1:4, ] + 5000
coord[5:8, ] <- coord[5:8, ] + 6000
corner <- rep(c(1, 2, 4, 3), 2)
plot <- rep(c("plot1", "plot2"), each = 4)

cut <- cutPlot(coord, plot, corner, gridsize = 100, dimX = 200, dimY = 200)


# Assign a plot to 200 trees
plot <- rep(c("plot1", "plot2"), 100)

# attribute trees to subplots
attributeTreeCoord(xy, plot, coordAbs = cut)


}
