% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheManager.R
\name{cacheManager}
\alias{cacheManager}
\title{Manage the cache}
\usage{
cacheManager(nameFile)
}
\arguments{
\item{nameFile}{the name of the file or folder}
}
\value{
the path to the file we need
}
\description{
The function creates a folder (only once) and then controls files that need to be downloaded and
placed in this folder.
}
\section{Localisation}{

The localisation of the folder is :
\itemize{
\item On Linux : \code{~/.local/share/BIOMASS}
\item On Mac OS X : \code{~/Library/Application Support/BIOMASS}
\item On Windows 7 up to 10 : \code{C:\\Users\\<username>\\AppData\\Local\\BIOMASS\\BIOMASS}
\item On Windows XP : \code{C:\\Documents and Settings\\<username>\\Data\\BIOMASS\\BIOMASS}
}

See this function for more information : \code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}}
}

\seealso{
\code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}}
}
\author{
Arthur PERE
}
\keyword{Internal}
