% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheManager.R
\name{cacheManager}
\alias{cacheManager}
\title{Manage the cache}
\usage{
cacheManager(nameFile, useCache = FALSE)
}
\arguments{
\item{nameFile}{the name of the file or folder}

\item{useCache}{logical. Whether or not use a cache.}
}
\value{
the path to the file of interest (character)
}
\description{
The function creates a folder (only once) and then controls files that need to be downloaded and
placed in this folder.
}
\section{Localisation}{

The localisation of the folder is :
\itemize{
\item On Linux : \verb{~/.local/share/R/BIOMASS}
\item On Mac OS X : \verb{~/Library/Application Support/R/BIOMASS}
\item On Windows 7 up to 10 : \verb{C:\\\\Users\\\\<username>\\\\AppData\\\\Local\\\\R\\\\BIOMASS\\\\R\\\\BIOMASS}
\item On Windows XP : \verb{C:\\\\Documents and Settings\\\\<username>\\\\Data\\\\R\\\\BIOMASS\\\\R\\\\BIOMASS}
}

See this function for more information : \code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}}
}

\seealso{
\code{\link[rappdirs:user_data_dir]{rappdirs::user_data_dir()}}
}
\author{
Arthur PERE
}
\keyword{Internal}
