% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_plot.R
\name{divide_plot}
\alias{divide_plot}
\title{Divides one ore more plots into subplots}
\usage{
divide_plot(
  rel_coord,
  proj_coord = NULL,
  grid_size,
  plot_ID_corner = NULL,
  tree_coord = NULL,
  plot_ID_tree = NULL
)
}
\arguments{
\item{rel_coord}{a data frame containing the relative (local) coordinates of plot corners, with X and Y on the first and second column respectively}

\item{proj_coord}{(optional) a data frame containing the projected coordinates of plot corners, with X and Y on the first and second column respectively, and with the same row order than rel_coord}

\item{grid_size}{a vector indicating the dimensions of grid cells (resp. X and Y dimensions). If only one value is given, the grid cells will be considered as squares.}

\item{plot_ID_corner}{if dealing with multiple plots : a vector indicating plot IDs for corners.}

\item{tree_coord}{(otpional) a data frame containing at least the relative tree coordinates (field/local coordinates), with X and Y corresponding to the first and second columns respectively}

\item{plot_ID_tree}{if dealing with multiple plots : a vector indicating tree plot IDs.}
}
\value{
Returns a data-frame containing as many rows as there are corners corresponding to the subplots, and the following columns :
\itemize{
\item \code{plot_ID_corner}: If dealing with multiple plots : the plot code
\item \code{subplot}: The automatically generated subplot code
\item \code{Xrel}:  The relative X-axis coordinates of subplots corners
\item \code{Yrel}:  The relative Y-axis coordinates of subplots corners
\item \code{X}:  If proj_coord is supplied : the projected X-axis coordinates of subplots corners
\item \code{Y}:  If proj_coord is supplied : the projected Y-axis coordinates of subplots corners
}
}
\description{
This function divides a plot (or several plots) into subplots in the relative coordinates system, and returns the coordinates of subplot corners
If corner coordinates in the projected coordinate system is supplied (proj_coord), subplot projected coordinates are calculated by a bilinear interpolation in relation with plot corner relative coordinates.
}
\examples{

coord <- data.frame(X = c(0, 200, 0, 200), Y = c(0, 0, 200, 200)) + 5000
cornerNum <- c(1, 2, 4, 3)
plot <- rep("plot1", 4)

cut <- cutPlot(coord, plot, cornerNum, grid_size = 100, dimX = 200, dimY = 200)

# plot the result
plot(coord, main = "example", xlim = c(4900, 5300), ylim = c(4900, 5300), asp = 1)
text(coord, labels = cornerNum, pos = 1)
points(cut$XAbs, cut$YAbs, pch = "+")
legend("bottomright", legend = c("orignal", "cut"), pch = c("o", "+"))

}
\author{
Arthur PERE, Arthur BAILLY
}
