% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{NouraguesCoords}
\alias{NouraguesCoords}
\title{Nouragues plot coordinates}
\format{
A data frame with 16 observations (GPS measurements) of the 8 following variables :
\itemize{
\item \code{Site}: Name of the site set up in the Nouragues forest
\item \code{Plot}: Plot ID of the site
\item \code{Xfield}: Corner location on the x-axis in the local coordinate system (defined by the 4 corners of the plot)
\item \code{Yfield}: Corner location on the y-axis in the local coordinate system
\item \code{Xutm}: Corner location on the x-axis in the UTM coordinate system
\item \code{Yutm}: Corner location on the y-axis in the UTM coordinate system
\item \code{Long}: Corner longitude coordinate
\item \code{Lat}: Corner latitude coordinate
}
}
\usage{
data(NouraguesCoords)
}
\description{
Dataset containing the corner coordinates of 4 plots of `Petit Plateau' in Nouragues forest (French Guiana).
}
\examples{
data(NouraguesCoords)
str(NouraguesCoords)
}
\references{
Jaouen, Gaëlle, 2023, "Nouragues forest permanent plots details", \doi{10.18167/DVN1/HXKS4E}, CIRAD Dataverse, V2
}
\keyword{datasets}
