\name{NouraguesHD}
\alias{NouraguesHD}
\docType{data}
\title{
Height-Diameter data
}
\description{
Dataset from two 1-ha plots from the Nouragues forest (French Guiana)
}
\usage{data("NouraguesHD")}
\format{
  A data frame with 1033 observations on the following 3 variables :
  \describe{
    \item{\code{plotId}}{Names of the plots}
	\item{\code{D}}{Diameter}
    \item{\code{H}}{Height}
  }
}
\examples{
data(NouraguesHD)
str(NouraguesHD)
}
\keyword{datasets}
