% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{param_4}
\alias{param_4}
\title{Posterior distribution of Chave et al.'s 2014 equation 4 parameters}
\format{
A data frame with 1001 observations on the following 3 variables.
\itemize{
\item \code{intercept}: Vector of intercept values
\item \code{logagbt}: Vector of the model coefficients associated with the product wood density * diameter^2 * height
\item \code{sd}: Vector of model residual standard error (RSE) values
}
}
\usage{
data("param_4")
}
\description{
This matrix contains the posterior distribution of the parameters of Equation 4 of Chave et al. (2014),
obtained in a Bayesian framework with uninformative priors through a Metropolis algorithm.
}
\details{
This dataset is used in the function \code{\link[=AGBmonteCarlo]{AGBmonteCarlo()}}.
}
\examples{
data(param_4)
str(param_4)
}
\references{
Chave et al. (2014) \emph{Improved allometric models to estimate the aboveground biomass of tropical trees},
Global Change Biology, 20 (10), 3177-3190
}
\keyword{AGBmonteCarlo}
\keyword{datasets}
\keyword{internal}
