% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrust.R
\name{procrust}
\alias{procrust}
\title{Procrust analysis}
\usage{
procrust(X, Y)
}
\arguments{
\item{X}{the target matrix}

\item{Y}{the matrix we want to fit to the target}
}
\value{
A list with the translation vector and the matrix of rotation
}
\description{
Do a procrust analysis. X is the target matrix, Y is the matrix we want to fit to the target.
This function returns a translation vector and a rotation matrix
After the procrust problem you \strong{must} do the rotation before the translation.
\strong{Warning :} The order of the between the lines of the matrix X and Y must be respected. #####??????
\strong{The order of the value on both matrix is important}
}
\author{
Arthur PERE
}
\keyword{Internal}
\keyword{analysis}
\keyword{procrust}
