% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot_summary.R
\name{subplot_summary}
\alias{subplot_summary}
\title{Summarise and display tree information by subplot}
\usage{
subplot_summary(subplots, value = NULL, draw_plot = TRUE, fun = sum, ...)
}
\arguments{
\item{subplots}{output of the \code{\link[=divide_plot]{divide_plot()}} function}

\item{value}{a character indicating the column in subplots$tree_data to display}

\item{draw_plot}{a logical indicating whether the plot design should be displayed and returned}

\item{fun}{the function to be applied}

\item{...}{optional arguments to fun}
}
\value{
a list containg the following elements :
\itemize{
\item \code{tree_summary} : a summary of the metric per subplot
\item \code{polygon} : an sf object : simple feature collection of the subplot's polygon
\item \code{plot_design} : (if draw_plot=T) a ggplot object (or a list of ggplot objects) that can easily be modified
}
}
\description{
After applying the \code{\link[=divide_plot]{divide_plot()}} function, this function summarises with any defined function the desired tree metric by sub-plot and displays the plot representation.
}
\examples{
rel_coord <- data.frame(x_rel = c(0, 200, 0, 200), y_rel = c(0, 0, 200, 200))
proj_coord <- data.frame(x_proj = c(210, 383, 110, 283), y_proj = c(210, 310, 383, 483))
tree_data <- data.frame(x_tree = runif(50,0,200), y_tree = runif(50,0,200), metric = rnorm(50,10,5))
subplots <- divide_plot(rel_coord, proj_coord = proj_coord, grid_size = 50,
                        tree_data = tree_data, tree_coords = c("x_tree","y_tree"))
# Sum summary (by default)
subplot_summary(subplots , value = "metric", draw_plot = FALSE)
# 9th quantile summary (for example)
subplot_summary(subplots , value = "metric", draw_plot = FALSE,
                fun = quantile, probs=0.9)


}
\author{
Arthur Bailly
}
