% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryByPlot.R
\name{summaryByPlot}
\alias{summaryByPlot}
\title{Summarize the matrix given by \code{\link{AGBmonteCarlo}} by plot (or subplot)}
\usage{
summaryByPlot(AGB_simu, plot, subplot = plot)
}
\arguments{
\item{AGB_simu}{Matrix which result for the function \code{\link{AGBmonteCarlo}} on the AGB_simu element of the list}

\item{plot}{Vector with the code of plot}

\item{subplot}{vector with the code of the subplot}
}
\value{
a data frame where :
\describe{
   \item{plot/subplot}{the code of the plot/subplot}
   \item{AGB}{the mean of AGB for the plot or subplot}
   \item{Cred_2.5}{the quantile 2.5\% for the plot or subplot}
   \item{Cred_97.5}{the quantile 97.5\% for the plot or subplot}
}
}
\description{
This function summarize the matrix \code{AGB_simu} given by the function \code{\link{AGBmonteCarlo}} by plot, or subplot 
if given. 
It's gives the mean, and the credibility intervalle [2.5\%, 97.5\%].

For the trees who haven't subplot because the coordinate of the trees isn't in the range of the plot 
(you can mark those by NA in the subplot vector), the AGB of those trees will be spread between the trees of the plot.
More precisly, the AGB will be randomly add to the first tree in each subplots for the given plot, for each iteration 
in the function \code{\link{AGBmonteCarlo}}.

For the trees who haven't any attribute plots (mark as NA in the plot vector), the same treatment than earlier will apply
expect for all the fist trees in all subplot (indepent of the plot).
}
\examples{

# Load a database
data(NouraguesHD)
data(KarnatakaForest)

# Modelling height-diameter relationship
HDmodel <- modelHD(D = NouraguesHD$D, H = NouraguesHD$H, method = "log2")

# Retrieving wood density values
KarnatakaWD <- getWoodDensity(KarnatakaForest$genus, KarnatakaForest$species,
  stand = KarnatakaForest$plotId
)

# Propagating errors with a standard error in wood density in one plot
filt <- KarnatakaForest$plotId \%in\% c("BSP20", "BSP14")
resultMC <- AGBmonteCarlo(
  D = KarnatakaForest$D[filt], WD = KarnatakaWD$meanWD[filt],
  errWD = KarnatakaWD$sdWD[filt], HDmodel = HDmodel
)

plot <- KarnatakaForest$plotId[ KarnatakaForest$plotId \%in\% c("BSP20", "BSP14") ]

# The summary by plot
summaryByPlot(AGB_simu = resultMC$AGB_simu, plot)
}
