% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryByPlot.R
\name{summaryByPlot}
\alias{summaryByPlot}
\title{Summarize by plot (or subplot) the posterior distribution of AGB values}
\usage{
summaryByPlot(AGB_simu, plot)
}
\arguments{
\item{AGB_simu}{Matrix resulting from the function \code{\link[=AGBmonteCarlo]{AGBmonteCarlo()}} (AGB_simu element of the list),
or just the result of the function \code{\link[=AGBmonteCarlo]{AGBmonteCarlo()}}. Or the vector resulting from the function \code{\link[=computeAGB]{computeAGB()}}}

\item{plot}{Vector with the code of plot}
}
\value{
a data frame where:
\itemize{
\item \code{plot}: the code of the plot
\item \code{AGB}: the mean of AGB for the plot
\item \code{Cred_2.5}: the quantile 2.5\% for the plot, if it's the matrix
\item \code{Cred_97.5}: the quantile 97.5\% for the plot, if it's the matrix
}
}
\description{
This function summarize the matrix \code{AGB_simu} given by the function \code{\link[=AGBmonteCarlo]{AGBmonteCarlo()}} by plot. Or just do the sums
for each plot of the AGB if the argument \code{AGB_simu} is the resulting vector from the function \code{\link[=computeAGB]{computeAGB()}}.
}
\details{
If some trees belong to an unknown plot (i.e. NA value in the plot arguments), their AGB values are randomly assigned
to a plot at each iteration of the AGB monte Carlo approach.
}
\examples{

# Load a database
data(NouraguesHD)
data(KarnatakaForest)

# Modelling height-diameter relationship
HDmodel <- modelHD(D = NouraguesHD$D, H = NouraguesHD$H, method = "log2")

# Retrieving wood density values
KarnatakaWD <- getWoodDensity(KarnatakaForest$genus, KarnatakaForest$species,
  stand = KarnatakaForest$plotId
)

# Propagating errors
filt <- KarnatakaForest$plotId \%in\% c("BSP20", "BSP14")
resultMC <- AGBmonteCarlo(
  D = KarnatakaForest$D[filt], WD = KarnatakaWD$meanWD[filt],
  errWD = KarnatakaWD$sdWD[filt], HDmodel = HDmodel
)

plot <- KarnatakaForest$plotId[ filt ]

# The summary by plot
summaryByPlot(AGB_simu = resultMC$AGB_simu, plot)
}
